Added new status', add TUSExecutor for networking, moved networking out of TUSClient, setup early example
