+++ Example/TUSKit/AppDelegate.swift
@@ -18,7 +18,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
         // Override point for customization after application launch.
         TUSClient.setup(with: TUSConfig(withUploadURLString: "https://master.tus.io/files"))
-        TUSClient.shared
         return true
     }
